Move conversion functions to a separate module
